+++ src/api/news/queries.ts
